Add a minimum blob size on the rolling checksum
For each blob, we record its mode, type, hash and offset, which use at least
57 bytes. The rolling checksum often generates blobs smaller than 57 bytes.
Not only those blobs end up using more space, but they also increase the
number of tree created to fanout the blobs. We also need to check if each
one is already present in the repository during the backup and we do more
disk seek during the restoration.

Of course, the minimum size also limit the minimum deduplication that can
be done, but, for size close to 57, the net result should be a gain in
speed and in space.

Signed-off-by: Patrick Rouleau <prouleau72@gmail.com>