Speedup saving by doing less
This patch is a proposition for a more optimistic behaviour of
PackWriter, which turns out to be effective when saving files that have
only changed a little.

Basically, instead of maybe_writing every blob on new_blob, we put them
in a queue. When the next tree will be written through new_tree, we
check if the tree already existed. If it did, then all its blobs too, so
we don't even need to test their existence. Otherwise we maybe_write
them.

This decision was driven by the status of one of my test repos,
consisting of dozens of middle-sized files (~MB) and a few
large-size files (~GB):

> cd $BUP_DIR/objects/pack
> for i in `ls *.idx`; do git verify-pack -v $i | cut -f 2 -d ' ' | grep -E '(blob|tree|commit)' ; done | sort | uniq -c

 312775 blob
     39 commit
  23324 tree

As you can see, there are 10 times more blobs than trees. By checking
only trees' existence, one could expect better performance.

My test was to `touch` a few large files (for a total of 1.9 GB) so
that no actual content was changed (it was already indexed in the
repo) but it was detected as new by `bup index` , and see the
difference with the previous version on a `bup save`. After a few
runs, I go down from 30s to 25s. There's also a saving on `bup split`,
  but not that big (a few seconds out of ~30).

Note that no tests were added, since this is speeding up the core feature
that is already well tested (and it passed `make test` successfully)

Unfortunately generalizing this behaviour to higher levels (queue
trees until we have a big enough queue to check, and recursively
discard objects if the parent is already there) turns out to be
worse in terms of performance.

Signed-off-by: Matthieu Rakotojaona <matthieu.r...@gmail.com>