bup-cron: allow the user to replace the "--strip" by "--graft"
By using "--graft", the directory structure is the same with or without a
snapshot. And since the user is using a jobfile, he can specify the name
he want and we do not have to decorate the branch name with the path anymore.

--strip itself is replaced by --strip-path because we may have more than
one include paths during the save, but we only want to strip the root part.

Signed-off-by: Patrick Rouleau <prouleau72@gmail.com>