options.py: Add a numeric option spec
When an option ends with a '=', it needs an argument, but before
this commit, the argument was automatically converted to an int,
which caused problems with some options that expected a string,
e.g. "bup-save -n 201502".

This commit add '+' as a new option spec. Now, when an option ends
with a '=', it needs a string argument, and when it ends with a
'+', it needs an integer argument.

Signed-off-by: Patrick Rouleau <prouleau72@gmail.com>